requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.7
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-disqus==0.5
django-embed-video==1.4.0
django-extensions==3.2.1
django-import-export==2.6.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-taggit==1.2.0
django-utils-six==2.0
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
mock==4.0.3
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.3.2
psycopg2-binary==2.9.1
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
urllib3==1.26.7
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0


